Lab 4 fixed

+++ src/AbstractFactory.java
@@ -1,8 +1,8 @@
 public interface AbstractFactory
 {
     Ordinary_Rabbit createOrdinary();
-    Ordinary_Rabbit createOrdinary(float x, float y, long time);
+    Ordinary_Rabbit createOrdinary(int x, int y, long time);
 
     Albino createAlbino();
-    Albino createAlbino(float x, float y, long time);
+    Albino createAlbino(int x, int y, long time);
 }
